% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkline_util.R
\name{add_vline}
\alias{add_vline}
\title{add_vline}
\usage{
add_vline(p, x, ...)
}
\arguments{
\item{p}{A parameter indicate which plot to draw the vertical line}

\item{x}{A number indicate where to add vertical line on the x_axis}
}
\description{
This function is a internal function drawing the vertical lines on the plot.
}
\examples{
p <- plot_ly(cars, x = ~speed, y = ~dist)
add_vline = function(p, x, ...) {
  l_shape = list(
    type = "line",
    y0 = 0, y1 = 1, yref = "paper",
    x0 = x, x1 = x,
    line = list(...)
 )
 p \%>\% layout(shapes=list(l_shape))
}
add_vline(p, x = 7.5)
}
