% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_summary_table.R
\name{ae_summary_table}
\alias{ae_summary_table}
\title{ae_summary_table}
\usage{
ae_summary_table(
  population_from,
  observation_from,
  population_where,
  observation_where,
  treatment_var,
  treatment_order,
  ae_var,
  ae_interested = NULL,
  stratum_var = NULL,
  display_ci = FALSE,
  display_total = FALSE,
  title_text,
  subtitle_text = NULL,
  end_notes,
  output_name = "./tlf_ae_summary.rtf"
)
}
\arguments{
\item{population_from}{A data frame to obtain population level variables.Typically an 'adsl' dataset.}

\item{observation_from}{A data frame to obtain observation level variables.}

\item{population_where}{A character string to define the criteria to select analysis population.}

\item{observation_where}{A character string to define the criteria to select analysis observation.}

\item{treatment_var}{A character string to define the variable of new column called "treatment"}

\item{treatment_order}{A character vector to define the treatment display order and label.}

\item{ae_var}{A character string to define the variable of new column called ae}

\item{ae_interested}{An object returned by function define_ae_select_list()}

\item{stratum_var}{A character string to define the variable of baseline stratum in 'population_from'.Only one 'stratum_var' is allowed.}

\item{display_ci}{A logic value indicating whether displaying the risk difference, corresponding confidence interval and its p-value. Default is FALSE.}

\item{display_total}{A logic value indicating whether displaying the total column. Default is FALSE.}

\item{title_text}{The part of the text string to appear in the title row.}

\item{subtitle_text}{A vector of text strings to appear in the subtitle row(s).}

\item{end_notes}{A vector of text strings to appear in the footnote.}

\item{output_name}{A text string to output the table, can be specified with the path together. If no path specified, the default path will be used to save the table.}
}
\value{
Return a table for summary of AE
}
\description{
This function is used to generate AE summary table
}
\examples{
library(dplyr)
library(r2rtf)
ae_summary_table(population_from = adsl \%>\% rename(TRTA = TRT01A),
                observation_from = adae,
                population_where = "ITTFL=='Y'",
                observation_where = NULL,
                treatment_var = "TRTA",
                treatment_order = c("MK9999" = "Xanomeline High Dose", "Placebo" = "Placebo"),
                ae_var = "AEDECOD",
                ae_interested = define_ae_select_list(ae_criterion = c('AESER == "Y"', 'AEREL != "NONE"',
                                                                       'AEACN == "DRUG WITHDRAWN"'),
                                                      ae_label = c("with serious adverse events",
                                                                   "with drug-related adverse events",
                                                                   "discontinued due to an adverse event")),
                stratum_var = NULL,
                display_ci = TRUE,
                display_total = FALSE,
                title_text = "Analysis of Adverse Event Summary", 
                subtitle_text = NULL,
                end_notes ="Every subject is counted a single time for each applicable row and column.",
                output_name = file.path(tempdir(), 'ae0summary.rtf'))
}
