% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_ae_selectList.R
\name{define_ae_selectList}
\alias{define_ae_selectList}
\title{To formate the interested ae and its filter
For example, if the interested ae are c("AESER", "AEREF"),
then their accoding filters can be defined as c("N", "None")
to filter OUT the NONE-SERIOUS ae and NONE-DRUG-RELATED AE}
\usage{
define_ae_selectList(ae_criterion = NULL, ae_label = NULL)
}
\arguments{
\item{ae_criterion}{a character vector to specify the AE criterion to filter the interested ones.
For example,
1) ae_criterion = c('AESER != "N"', 'AEREL != "None"')
2) ae_criterion = c('AESER != "N"', 'AEREL != "None"', 'Grade3To5Flag != "Y"', 'AEACN != "DRUG WITHDRAWN"')}

\item{ae_label}{a string vector to specify the according AE labels to be displayed in the selected list.
The default value is c("with serious adverse events", "with drug-related adverse events",
"with toxicity grade 3-5 adverse events","discontinued due to an adverse event")}
}
\value{
a list including
(1) the ae filters
(3) the ae labels
}
\description{
To formate the interested ae and its filter
For example, if the interested ae are c("AESER", "AEREF"),
then their accoding filters can be defined as c("N", "None")
to filter OUT the NONE-SERIOUS ae and NONE-DRUG-RELATED AE
}
\examples{
define_ae_selectList(ae_criterion = 'AESER != "N"', ae_label = "with serious adverse events")
define_ae_selectList(ae_criterion = c('AESER != "N"', 'AEREL != "None"'),
                     ae_label = c("with serious adverse events", "with drug-related adverse events"))

}
