% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ae.R
\name{download_ae}
\alias{download_ae}
\title{Generate download button for ae tables}
\usage{
download_ae(db, ae_label_download, source = NULL, footnote = NULL)
}
\arguments{
\item{db}{ae data frame usually generated by tidy_ae_table2(). It should contain the following columns:
(1) ae: the name of ae
(2) ae_label: whether no filter, or filter with AESER, or filter with AEREL...
(3) n_1: the number of unique ae subjects in the treatment arm
(4) n_2: the number of unique ae subjects in the control arm
(5) pct_1: the percentage of unique ae subjects in the treatment arm
(6) pct_2: the percentage of unique ae subjects in the coltrol arm}

\item{ae_label_download}{a string vector, specifying the ae label to be download.
For example, it can be "with serious adverse events",
or c("with serious adverse events", "with drug-related adverse events")}

\item{source}{a string to specify the text of source in the downloaded ae summary tables}

\item{footnote}{a string to specify the text of footnote in the downloaded ae summary tables}
}
\value{
some download buttons
}
\description{
Generate download button for ae tables
}
\examples{
\dontrun{
tbl <- data.frame(ae = c("headache", "pain", "fever", "running nose", "fever", "headache", "running nose"),
                  ae_label = c("All", "All", "All", "All", "AESER", "AEREL", "AEREL"),
                  n_1 = c(2, 3, 1, 5, 1, 1, 3),
                  n_2 = c(4, 5, 3, 7, 3, 3, 5),
                  N_1 = rep(20, 7),
                  N_2 = rep(20, 7),
                  stratum = rep("NULL", 7))
tbl <- tbl \%>\% dplyr::mutate(pct_1 = n_1/N_1, pct_2 = n_2/N_2)

treatment_order = c("MK9999" = "Xanomeline", "Placebo" = "Placebo")

db <- list(table = tbl, treatment_order = treatment_order)
download_ae(db, ae_label_download = c("AESER", "AEREL"))
}
}
