% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{forestly-package}
\alias{forestly}
\alias{forestly-package}
\title{forestly: Interactive Forest Plot}
\description{
Interactive Forest Plot.
}
\author{
\strong{Maintainer}: Yilong Zhang \email{yilong.zhang@merck.com}

}
\keyword{internal}
