% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{forestly_adae_3grp}
\alias{forestly_adae_3grp}
\title{An adverse event dataset}
\format{
A data frame with 1191 rows and 56 variables.
}
\source{
\url{https://github.com/phuse-org/phuse-scripts/tree/master/data/adam/cdisc}
}
\usage{
forestly_adae_3grp
}
\description{
A dataset containing the adverse event information of a clinical trial
following the CDISC ADaM standard.
}
\details{
Definition of each variable can be found at
\url{https://github.com/phuse-org/phuse-scripts/tree/master/data/adam/cdisc}.
}
\keyword{datasets}
