% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{forestly_adsl}
\alias{forestly_adsl}
\title{A subject level demographic dataset}
\format{
A data frame with 170 rows and 49 variables.
}
\source{
\url{https://github.com/phuse-org/phuse-scripts/tree/master/data/adam/cdisc}
}
\usage{
forestly_adsl
}
\description{
A dataset containing the demographic information of a clinical trial
following the CDISC ADaM standard.
}
\details{
Definition of each variable can be found at
\url{https://github.com/phuse-org/phuse-scripts/tree/master/data/adam/cdisc}.
}
\keyword{datasets}
