% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_ae_forestly.R
\name{format_ae_forestly}
\alias{format_ae_forestly}
\title{Format outdata for interactive forest plot}
\usage{
format_ae_forestly(
  outdata,
  display = c("n", "prop", "fig_prop", "fig_diff"),
  digits = 1,
  width_term = 200,
  width_fig = 320,
  width_n = 40,
  width_prop = 60,
  width_diff = 80,
  footer_space = 90,
  prop_range = NULL,
  diff_range = NULL,
  color = NULL,
  ae_col_header = NULL,
  diff_label = "Treatment <- Favor -> Placebo",
  diff_col_header = NULL,
  diff_fig_header = NULL
)
}
\arguments{
\item{outdata}{An \code{outdata} object created by \code{\link[=prepare_ae_forestly]{prepare_ae_forestly()}}.}

\item{display}{A character vector of measurement to be displayed.
\itemize{
\item \code{n}: Number of subjects with AE.
\item \code{prop}: Proportion of subjects with AE.
\item \code{total}: Total columns.
\item \code{diff}: Risk difference.
}}

\item{digits}{A number of digits after decimal point to be displayed for proportion and
risk difference.}

\item{width_term}{Width in px for AE term column.}

\item{width_fig}{Width in px for proportion and risk difference figure.}

\item{width_n}{Width in px for "N" columns.}

\item{width_prop}{Width in px for "(\%)" columns.}

\item{width_diff}{Width in px for risk difference columns.}

\item{footer_space}{Space in px for footer to display legend.}

\item{prop_range}{A vector of lower and upper limit of x-axis
for proportion figure.}

\item{diff_range}{A vector of lower and upper limit of x-axis
for risk difference figure.}

\item{color}{A vector of colors for analysis groups.
Default value supports up to 4 groups.}

\item{ae_col_header}{Column header for adverse events item columns.
If NULL (default) and "par" specified in \code{components} from \code{prepare_ae_forestly()}, uses "Adverse Event".
If NULL and "soc" specified in \code{components} from \code{prepare_ae_forestly()}, uses "System Organ Class" for "soc".}

\item{diff_label}{x-axis label for risk difference.}

\item{diff_col_header}{Column header for risk difference table columns.
If NULL (default), uses "Risk Difference (\%) \if{html}{\out{<br>}} vs. Reference Group".}

\item{diff_fig_header}{Column header for risk difference figure.
If NULL (default), uses "Risk Difference (\%) + 95\% CI \if{html}{\out{<br>}} vs. Reference Group".}
}
\value{
An \code{outdata} object.
}
\description{
Format outdata for interactive forest plot
}
\examples{
adsl <- forestly_adsl
adae <- forestly_adae
adsl$TRTA <- factor(
  adsl$TRT01A,
  levels = c("Xanomeline Low Dose", "Placebo"),
  labels = c("Low Dose", "Placebo")
)
adae$TRTA <- factor(
  adae$TRTA,
  levels = c("Xanomeline Low Dose", "Placebo"),
  labels = c("Low Dose", "Placebo")
)

analysis_plan <- metalite::plan(
  analysis = "ae_forestly",
  population = "apat",
  observation = "wk12",
  parameter = "any"
)
meta <- metalite::meta_adam(population = adsl, observation = adae) |>
  metalite::define_plan(plan = analysis_plan) |>
  metalite::define_population(
    name = "apat",
    var = c("USUBJID", "SAFFL", "TRTA", "SITEID", "SEX", "RACE", "AGE"),
    group = "TRTA",
    subset = SAFFL == "Y",
    label = "All Participants as Treated"
  ) |>
  metalite::define_observation(
    name = "wk12",
    var = c(
      "USUBJID", "SAFFL", "TRTA", "SITEID", "SEX", "RACE", "AGE",
      "ASTDY", "AEDECOD", "AEBODSYS", "AESER", "AEREL", "AEACN",
      "AEOUT", "ADURN", "ADURU"
    ),
    group = "TRTA",
    subset = SAFFL == "Y",
    label = "Weeks 0 to 12"
  ) |>
  metalite::define_parameter(
    name = "any",
    term1 = "",
    term2 = "",
    var = "AEDECOD",
    soc = "AEBODSYS",
    label = "All AEs"
  ) |>
  metalite::define_analysis(
    name = "ae_forestly",
    label = "Interactive forest plot"
  ) |>
  metalite::meta_build()

meta |>
  prepare_ae_forestly(parameter = "any") |>
  format_ae_forestly()
}
