% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_reactable.R
\name{mk_reactable}
\alias{mk_reactable}
\title{mk_reactable}
\usage{
mk_reactable(
  data,
  resizable = TRUE,
  filterable = TRUE,
  searchable = TRUE,
  defaultPageSize = 10,
  borderless = TRUE,
  striped = TRUE,
  highlight = TRUE,
  ...
)
}
\description{
mk_reactable
}
\examples{

mk_reactable(iris)

}
