% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{nudge_unit}
\alias{nudge_unit}
\title{Nudge around a unit}
\usage{
nudge_unit(n)
}
\arguments{
\item{n}{A number of a group within a unit}
}
\value{
A nudge for unit
}
\description{
\code{nudge_unit} nudges around a unit
}
\examples{
nudge_unit(10)
}
