% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_compare.R
\name{rate_compare}
\alias{rate_compare}
\title{Unstratified and Stratified  Miettinen and Nurminen Test}
\usage{
rate_compare(
  formula,
  data,
  delta = 0,
  weight_schema = c("ss", "equal", "cmh"),
  test = c("one.sided", "two.sided"),
  bisection = 100,
  eps = 1e-06,
  alpha = 0.05
)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted, which has the form \code{y ~ x} for unstratified MN test, and the form \code{y ~ x + strata(s)} for stratified MN test, where \code{y} is the numeric vector with values of 0 or 1, \code{x} is the group information, and \code{s} is the stratum information.}

\item{data}{an optional data frame, list or environment containing the variables in the model. If not found in data, the variables are taken from \code{environment (formula)}, typically the environment from which \code{rate_compare} is called.}

\item{delta}{a numeric value to set the difference of two group under the null.}

\item{weight_schema}{weighting schema used in stratified MN method. \code{"equal"} for equal weighting, \code{"ss"} for sample size weighting, \code{"cmh"} for Cochran Mantel-Haenszel's weights. Default is "ss".}

\item{test}{a character string specifying the side of p-value, must be one of \code{"one.sided"}, or \code{"two.sided"}.}

\item{bisection}{the number of sections in the interval used in Bisection Method. Default is 100.}

\item{eps}{the level of precision. Default is eps=1e-06.}

\item{alpha}{pre-difined alpha level for Confidence Interval}
}
\description{
Unstratified and Stratified  Miettinen and Nurminen Test
}
\examples{
##To conduct the stratified MN analysis with sample size weights:
treatment <- c(rep('pbo',100),rep('exp',100))
response <- c(rep(0,80),rep(1,20),rep(0,40),rep(1,60))
stratum <- c(rep(1:4,12),1,3,3,1,rep(1:4,12),rep(1:4,25))
rate_compare(formula=response~factor(treatment,levels=c('pbo','exp'))+strata(stratum),delta = 0, weight_schema='ss',test = 'one.sided',alpha=0.05)
}
\references{
Miettinen, O. and Nurminen, M, \emph{Comparative Analysis of Two Rates}. STATISTICS IN MEDICINE, 4:213-226, 1985.
}
