% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the forestly Shiny app}
\usage{
run_app(...)
}
\arguments{
\item{...}{Options passed to the golem app.}
}
\value{
A Shiny application object.
}
\description{
Run the forestly Shiny app
}
