% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkline_draw_axis.R
\name{sparkline_draw_axis}
\alias{sparkline_draw_axis}
\title{sparkline_draw_axis}
\usage{
sparkline_draw_axis(
  color,
  label,
  xlim,
  xlab = NULL,
  mode = "markers",
  showlegend = TRUE,
  height = ifelse(showlegend, 60, 30),
  margin_bottom = ifelse(showlegend, 0, height)
)
}
\arguments{
\item{color}{A logical vector used to determine the color of the axis}

\item{label}{A vector of names, when you layout the legend it will name the variables as what you put in label.}

\item{xlim}{A vector to set the limits of x-axis}

\item{xlab}{A label of the x-axis}

\item{mode}{A vector of shapes. Add lines or markers in the legend}

\item{showlegend}{Logical value to show legend.}

\item{height}{A number of the distance to determine the height of x-axis}

\item{margin_bottom}{Numeric value of bottom margin}
}
\value{
A x-axis with annotation on the bottom
}
\description{
sparkline_draw_axis
}
\examples{
library(ggplot2)
library(plotly)
color <- c("red", "blue")
label <- c("MK9999", "Placebo")
xlim  <- c(-2, 2)
xlab  <- "Example"

sparkline_draw_axis(color, label, xlim,
                    showlegend = FALSE, height = 30)

sparkline_draw_axis(color, label, xlim,
                    showlegend = TRUE, height = 60)

sparkline_draw_axis(color, label, xlim, xlab = "Example",
                    showlegend = FALSE, height = 60)

sparkline_draw_axis(color, label, xlim, xlab = "Example",
                    showlegend = TRUE, height = 100)

sparkline_draw_axis(color, label, xlim, xlab = "Example",
                    mode = "lines", showlegend = TRUE, height = 100)

sparkline_draw_axis(color, label, xlim, xlab = "Example",
                    mode = "markers+lines", showlegend = TRUE, height = 100)

}
