% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkline_errorbar.R
\name{sparkline_errorbar}
\alias{sparkline_errorbar}
\title{sparkline_errorbar}
\usage{
sparkline_errorbar(
  est,
  lower,
  upper,
  color = NULL,
  height = 30,
  xlim = NULL,
  vline = NULL,
  margin = list(l = 0, r = 0, b = 0, t = 0, pad = 0),
  ...
)
}
\arguments{
\item{est}{A vector  determined the point of your bar}

\item{lower}{A vector determined the leftmost point}

\item{upper}{A vector determined the rightmost point}

\item{color}{A vector of colors you want your bar to be}

\item{height}{A variable determines the distance between your bars}
}
\value{
Sparkline error bar for interactive graphics
}
\description{
This function is used to draw error bars based on the parameters you defined.
}
\examples{
library(plotly)
est <- c(0,1,2)
lower <- c(-1, -2, -1)
upper <- c(3, 5, 5)
color <- c("red", "blue", "gold")
sparkline_errorbar(est, lower, upper, color)

}
