% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkline_util.R
\name{sparkline_layout}
\alias{sparkline_layout}
\title{sparkline_layout}
\usage{
sparkline_layout(
  p,
  xlim = NULL,
  vline = NULL,
  color = "#00000050",
  margin = list(l = 0, r = 0, b = 0, t = 0, pad = 0),
  fixedrange = TRUE
)
}
\arguments{
\item{p}{A parameter indicate which plot to add color for scatters.}

\item{xlim}{A data frame define the range of x-axis on the plot}

\item{vline}{A data frame indicates where to draw the vertical lines on the plot}

\item{color}{A character string define the color of vertical line.}

\item{margin}{A data define the size of the plot}
}
\description{
This is the function which adding color to vertical line.
}
\examples{
p <- plot_ly(cars, x = ~speed, y = ~dist)
sparkline_layout <- function(p,
                             xlim  = NULL,
                             vline = NULL,
                             color = "#00000050",
                             margin = list(l = 0, r = 0, b = 0, t = 0, pad = 0),
                             fixedrange = TRUE
){

  stopifnot("plotly" \%in\% class(p))

  # Add vertical reference line
  l_shape = list(
    type = "line",
    y0 = 0, y1 = 1, yref = "paper", # i.e. y as a proportion of visible region
    x0 = vline,
    x1 = vline,
    line = list(color = color)
  )

  plotly::layout(p,
                 xaxis = list(title = "", range = xlim, zeroline = FALSE, fixedrange = fixedrange),
                 yaxis = list(title = "", showgrid = FALSE, zeroline = FALSE, showticklabels = FALSE, fixedrange = fixedrange),
                 margin = margin,
                 shapes = list(l_shape),
                 plot_bgcolor  = "rgba(0, 0, 0, 0)",
                 paper_bgcolor = "rgba(0, 0, 0, 0)",
                 hoverlabel=list(bgcolor="lightgray"),
                 showlegend = FALSE) \%>\%
    plotly::config(displayModeBar = FALSE)
}
sparkline_layout(p, color='red')

}
