% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkline_util.R
\name{sparkline_legend}
\alias{sparkline_legend}
\title{sparkline_legend}
\usage{
sparkline_legend(p, title = "Treatment", pos = -1)
}
\arguments{
\item{p}{A parameter indicate which plot to add a tittle}


}
\description{
This function is a internal function adding the title to the plot.
}
\examples{

}
