% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkline_point.R
\name{sparkline_point}
\alias{sparkline_point}
\title{sparkline_point}
\usage{
sparkline_point(
  x,
  y,
  color = NULL,
  height = 30,
  width = 150,
  hover_text = NULL,
  xlim = NULL,
  vline = NULL,
  margin = list(l = 0, r = 0, b = 0, t = 0, pad = 0),
  ...
)
}
\arguments{
\item{x}{a data frame corresponding location of scatters on x-axis}

\item{y}{a data frame corresponding location of scatters on y-axis}

\item{color}{A character string define the color of different scatters.}

\item{height}{Numeric value of figure height.}

\item{width}{Numeric value of figure width.}

\item{hover_text}{A character string describe the information of each scatters}

\item{xlim}{A data frame define the range of x-axis on the plot}

\item{vline}{A data frame indicates where to draw the vertical lines on the plot}

\item{margin}{A data define the size of the plot}

\item{...}{Additional arguments transfer to \code{sparkline_layout}}
}
\description{
This function is drawing the scatters on the sparkline plot
}
\examples{
library(plotly)
x <- 1:5
y <- 1:5
color <- c("red", "blue", "gold", "grey", "black")
sparkline_point(x = x, y = y, color = color, height = 500, vline = 1.5,
               hover_text = c('a','b','c','d','e'))
}
