% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkline_to_js.R
\name{sparkline_point_to_js}
\alias{sparkline_point_to_js}
\title{Transfor Plotly Sparkline Plot to Plotly Javascript Function for Reactable}
\usage{
sparkline_point_to_js(p, template = "inst/javascripts/sparkline_point.js")
}
\arguments{
\item{p}{a plotly object}

\item{template}{template javascript location}
}
\description{
Transfor Plotly Sparkline Plot to Plotly Javascript Function for Reactable
}
\keyword{internal}
