% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{theme_panel}
\alias{theme_panel}
\title{Theme function for plot with multiple panels}
\usage{
theme_panel(show_text = TRUE, show_ticks = TRUE)
}
\arguments{
\item{show_text}{A logical value that controls text display on the y axis.
Default is \code{TRUE}.}

\item{show_ticks}{A logical value that controls ticks display on the y axis.
Default is \code{TRUE}.}
}
\value{
Theme for a specific panel.
}
\description{
Specifies theme for a plot with multiple panels.
}
\examples{
library(ggplot2)

p <- ggplot(mpg, aes(displ, hwy, colour = class)) +
  geom_point()

p
p + theme_panel()
}
