% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_ae_listing.R
\name{tidy_ae_listing}
\alias{tidy_ae_listing}
\title{Output a data frame with the listed details in the reactable}
\usage{
tidy_ae_listing(db, listing_var = NULL, listing_label = NULL)
}
\arguments{
\item{db}{the data frame, whose subset will servers as the listed details in the reactable}

\item{listing_var}{a character vector specifying the variable names to be displayed in the listed details}

\item{listing_label}{a character vector specifying the variable labels to be displayed in the listed details}
}
\value{
a data frame with selected/renamed variables
}
\description{
Output a data frame with the listed details in the reactable
}
\examples{
library(tibble)
db = tibble(USUBJID = c("01", "01", "02", "03", "03", "03", "04", "04", "05"),
            SITEID = c("site1", "site1", "site2", "site3", "site3", "site3", "site4", "site4", "site5"),
            SEX = c("F", "F", "M", "F", "F", "F", "M", "M", "F"),
            AGE = c(65, 65, 30, 22, 22, 22, 70, 70, 53),
            RACE = c("Asian", "Asian", "Latino or Hispanic",
                     "WHITE", "WHITE", "WHITE",
                     "BLACK OR AFRICAN AMERICAN",
                     "BLACK OR AFRICAN AMERICAN",
                     "AMERICAN INDIAN OR ALASKA NATIVE"),
            ADURN = c(11, 3, 4, 5, 7, 20, 90, 34, 58),
            ADURU = rep("minutes", 9))

detail1 <- tidy_ae_listing(db, listing_var = c("USUBJID", "SITEID", "SEX", "RACE", "AGE"),
                           listing_label = c("ID", "Site Number", "Sex", "Race", "Age"))
detail2 <- tidy_ae_listing(db, listing_var = c("USUBJID", "SEX", "ADURN", "ADURU"),
                           listing_label = c("ID", "Gender", "Duration", "Duration"))
detail3 <- tidy_ae_listing(db, listing_var = c("USUBJID", "ADURN", "ADURU", "AGE"),
                           listing_label = c("ID", "Duration", "Duration", "Age"))
}
