% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_ae_table.R
\name{tidy_ae_table}
\alias{tidy_ae_table}
\title{tidy_ae_table}
\usage{
tidy_ae_table(
  population_from,
  observation_from,
  population_where = "ITTFL=='Y'",
  observation_where = NULL,
  treatment_var = "TRTA",
  treatment_order = unique(population_from[[treatment_var]]),
  ae_var = "AEDECOD",
  stratum_var = NULL,
  listing_var = names(observation_from)
)
}
\arguments{
\item{listing_var}{}
}
\value{
a table of ae table
}
\description{
get the whole ae table
}
\examples{

library(dplyr)
library(tidyr)
db <- adsl \%>\% rename(TRTA = TRT01A)
tidy_ae_table(population_from  = db,
              observation_from = adae,
              treatment_var = "TRTA",
              treatment_order = c("MK9999" = "Xanomeline High Dose", "Placebo" = "Placebo"))

}
