% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_ae_table1.R
\name{tidy_ae_table1}
\alias{tidy_ae_table1}
\title{tidy_ae_table1}
\usage{
tidy_ae_table1(
  population_from,
  observation_from,
  population_where = "ITTFL=='Y'",
  observation_where = NULL,
  treatment_var = "TRTA",
  treatment_order = unique(population_from[[treatment_var]]),
  ae_var = ae_var,
  ae_interested = ae_interested,
  stratum_var = NULL,
  listing_var = names(observation_from)
)
}
\arguments{
\item{population_from}{A data frame to obtain population level variables.Typically an 'adsl' dataset.}

\item{observation_from}{A data frame to obtain observation level variables.}

\item{population_where}{A character string to define the criteria to select analysis population.}

\item{observation_where}{A character string to define the criteria to select analysis observation.}

\item{treatment_var}{A character string to define the variable of new column called "treatment"}

\item{treatment_order}{A character vector to define the treatment display order and label.}

\item{ae_var}{A character string to define the variable of new column called ae}

\item{ae_interested}{A character string including the interested AE structure, ex., AEDECOD, AESOC, AESER, AEREL, AEOUT}

\item{stratum_var}{A character string to define the variable of baseline stratum in 'population_from'.Only one 'stratum_var' is allowed.}

\item{listing_var}{A character string to define the criteria to select the column of the table}
}
\value{
Return a standard adverse event data frame
}
\description{
Define analysis data with ONE AE structure,
for example, the only AE structure is AESER (serious events)
i.e., we have two boxes, depending on the number of sub-structure of AESER:
[]Y []N
This function is used to obtain AE information ready for visualization
}
\examples{
db <- tidy_ae_table1(population_from  = adsl_mk1293,
                    population_where = NULL,
                    observation_from = adae_mk1293,
                    observation_where = NULL,
                    treatment_var = "ACTLARM",
                    treatment_order = c("MK1293" = "MK-1293", "Placebo" = "Lantus"),
                    ae_var = "AEDECOD",
                    ae_interested = "AESER",
                    listing_var = c("SITEID", "USUBJID", "RACE", "SEX", "AESER", "AEREL", "AEACN", "AEOUT",
                                    "AESLIFE", "AESDTH", "AESOD", "AESMIE") )
}
