% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_multi_ae_label.R
\name{tidy_multi_ae_label}
\alias{tidy_multi_ae_label}
\title{This is a function to rbind data together according to the interested ae labels}
\usage{
tidy_multi_ae_label(db, db_N, ae_interested)
}
\arguments{
\item{db}{a data frame which is usually an intersection of adae and adsl,
with the ae variable name (usually encoded as "AEDECOD") transformed into "ae",
and treatment variable name (usually encoded as "TRTA") transformed into "treatment",
see the following example for the format of db}

\item{db_N}{a data frame contains the sample size of each arm}

\item{ae_interested}{A character string including the interested AE structure, ex., AESER, AEREL...}
}
\value{
a data frame which rbind data together according to the interested ae labels
}
\description{
This is a function to rbind data together according to the interested ae labels
}
\examples{
\dontrun{
library(dplyr)
treatment_order = c("MK9999" = "Xanomeline", "Placebo" = "Placebo")

db = tibble::tibble(USUBJID = c("01", "01", "02", "03", "03", "03", "04", "04", "05"),
                    treatment = factor(c("Xanomeline", "Xanomeline", "Placebo", "Placebo", "Placebo", 
                                          "Placebo", "Xanomeline", "Xanomeline", "Xanomeline"), 
                                       levels = treatment_order, labels = treatment_order),
                    ae = c("headache", "pain", "headache", "fever", "running nose", 
                           "pain", "headache", "fever", "headache"),
                    AESER = c("N", "N", "N", "Y", "N", "N", "N", "Y", "N"),
                    AEREL = c("N", "Y", "N", "N", "N", "Y", "N", "N", "Y"))
                    
db_N = tibble::tibble(treatment = factor(c("Xanomeline", "Placebo"), 
                                         levels = treatment_order, 
                                         labels = treatment_order), 
                      N = c(3, 2))

tidy_multi_ae_label(db, db_N, ae_interested = ae_interested(ae_criterion = c('AESER == "Y"', 'AEREL != "N"'),
                                                            ae_label = c("with serious adverse events","with drug-related adverse events")))
}
}
