% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_population1.R
\name{tidy_population1}
\alias{tidy_population1}
\title{tidy_population1}
\usage{
tidy_population1(
  population_from,
  population_where = NULL,
  treatment_var = "TRT01A",
  treatment_order = unique(population_from[[treatment_var]]),
  stratum_var = NULL,
  baseline_var = NULL
)
}
\arguments{
\item{population_from}{A data frame to obtain population level variables.
Typically an 'adsl' dataset.
It is the source of variables mentioned in
\code{population_where}, \code{treatment_var}, \code{stratum_var} and \code{covariate_var}.}

\item{population_where}{A character string to define the criteria to select analysis population.}

\item{treatment_var}{A character string to define the variable of new column called "treatment".}

\item{treatment_order}{A vector of character strings that tells the
function which rows of the table should be select,
only if the values in "treatment" is in this vector.
It also provide the label names of the treatments after turning them into factors.}

\item{stratum_var}{A character string to define the variable of baseline stratum in 'population_from'.
Only one 'stratum_var' is allowed.}

\item{baseline_var}{Tells the function which column of the table to be selected as the final output.}
}
\value{
an output data frame of population level information.
USUBJID: Unique subject identifier.
treatment: Actual treatment.
stratum: baseline stratum in 'population_from'.
}
\description{
This function is used to prepare the observation datasets for the use in function \code{tidy_ae_table()}.
The function filter the \code{adsl} dataset, and define proper treatment order in the dataset.
}
\examples{
pop <- tidy_population(population_from  = adsl \%>\% rename(TRTA = TRT01A),
 population_where = "ITTFL=='Y'",
 treatment_var    = "TRTA",
 treatment_order  = c("MK9999" = "Xanomeline High Dose", "Placebo" = "Placebo"),
 stratum_var      = NULL,
 baseline_var     = NULL)

tidy_population(population_from = adsl)
tidy_population(population_from = adsl, stratum_var = "RACE")
}
