% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlf_ae_specific.R
\name{tlf_ae_specific}
\alias{tlf_ae_specific}
\title{Specific Adverse Events Table}
\usage{
tlf_ae_specific(
  population_from,
  observation_from,
  population_where,
  observation_where,
  treatment_var,
  treatment_order,
  ae_var,
  ae_grp = NULL,
  display_total = FALSE,
  display_ci = FALSE,
  display_pval = FALSE,
  stratum_var = NULL,
  title_text,
  subtitle_text,
  end_notes,
  output_report = file.path(tempdir(), "ae0specific.rtf"),
  output_dataframe = file.path(tempdir(), "ae0specific.RData")
)
}
\arguments{
\item{population_from}{A data frame to obtain population level variables. Typically an 'adsl' dataset.}

\item{observation_from}{A data frame to obtain observation level variables.}

\item{population_where}{A character string to define the criteria to select analysis population.}

\item{observation_where}{A character string to define the criteria to select analysis observation.}

\item{treatment_var}{A character string to define the variable of new column called "treatment".}

\item{treatment_order}{A character vector to define the treatment display order and label.}

\item{ae_var}{A character string to define the variable of new column called aedecod.}

\item{ae_grp}{A character string to define the variable of new column called aebodsys, which is used to group AE terms. Default is NULL, for which no AE grouping is performed.}

\item{display_total}{A logic value indicating whether displaying the total column. Default is FALSE.}

\item{display_ci}{A logic value indicating whether displaying the risk difference and its corresponding confidence interval. Default is FALSE.}

\item{display_pval}{A logic value indicating whether displaying p-value for the risk difference. Default is FALSE.}

\item{stratum_var}{A character string to define the variable of baseline stratum in 'population_from'. Only one 'stratum_var' is allowed.
Default is NULL, for which nonstratified MN analysis will be conducted. If specified, stratified MN analysis will be conducted.}

\item{title_text}{The text string to appear in the title row.}

\item{subtitle_text}{A vector of text strings to appear in the subtitle row(s).}

\item{end_notes}{A vector of text strings to appear in the footnote.}

\item{output_report}{A text string to output the rtf table, can be specified with the path together.
If no path specified, the default path will be used to save the table.}

\item{output_dataframe}{A text string to output the rtf R object, like "./ae_summary.Rdata".
It can be specified with the path together. If no path specified, the default path will be used to save the table.}
}
\description{
The function creates specific adverse events tables by AE term or grouped by AE SOC. The user can select whether displaying
Total, confidence interval, or P-value columns.
}
\examples{
library(dplyr)
temp_dir <- tempdir()
tlf_ae_specific(population_from = adsl \%>\% rename(TRTA=TRT01A),
observation_from = adae,
population_where = "ITTFL == 'Y'",
observation_where = "TRTEMFL == 'Y'",
treatment_var = "TRTA",
treatment_order =c("MK9999" = "Xanomeline High Dose", "Placebo" = "Placebo"),
ae_var = "AEDECOD",
ae_grp = "AEBODSYS",
display_ci = FALSE,
display_total = FALSE,
display_pval = FALSE,
stratum_var = NULL,
title_text = "Participants With Adverse Events by System Organ Class and Preferred Term",
subtitle_text = c("(Incidence > 0\% in More or More Treatment Group)","(APaT Population)"),
end_notes = "Every subject is counted a single time for each applicable row and column.
Database Cutoff Date: 01SEP2021.",
output_report = file.path(temp_dir, 'ae0specific.rtf'),
output_dataframe = file.path(temp_dir, 'ae0specific.RData'))
}
