% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlf_download.R
\name{tlf_download}
\alias{tlf_download}
\title{Title}
\usage{
tlf_download(path, output_name = "AE TLF", button_label = "Download AE TLF")
}
\arguments{
\item{path}{The path where the tlf generated tlf_ae_summary or tlf_ae_specific is stored}

\item{output_name}{The folder name of the downloaded zip file}

\item{button_label}{The name to be displayed on the download button}
}
\value{
a downloaded zip folder
}
\description{
Title
}
\examples{
temp_dir <- tempdir()
tlf_ae_summary(population_from = adsl \%>\% rename(TRTA = TRT01A),
               observation_from = adae,
               population_where = "ITTFL=='Y'",
               observation_where = NULL,
               treatment_var = "TRTA",
               treatment_order = c("MK9999" = "Xanomeline High Dose", "Placebo" = "Placebo"),
               ae_var = "AEDECOD",
               ae_interested = define_ae_select_list(ae_criterion = c('AESER == "Y"', 'AEREL != "NONE"'),
                                                     ae_label = c("with serious adverse events",
                                                                  "with drug-related adverse events")),
               stratum_var = "STRATUMN", 
               display_ci = TRUE,
               display_total = FALSE,
               title_text = "Analysis of Adverse Event Summary",
               subtitle_text = NULL,
               end_notes = "Every subject is counted a single time for each applicable row and column.",
               output_name = file.path(temp_dir, 'ae0summary.rtf'))
tlf_download(path = temp_dir,
             output_name = "AE summary table", 
             button_label = "Download AE summary table")
}
