% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_reactable.R
\name{mk_reactable}
\alias{mk_reactable}
\title{Define default behavior of reactable
This function inputs a dataframe of observation for data,it will return data table use
reactable() on a data frame.}
\usage{
mk_reactable(
  data,
  resizable = TRUE,
  filterable = TRUE,
  searchable = TRUE,
  defaultPageSize = 10,
  borderless = TRUE,
  striped = TRUE,
  highlight = TRUE,
  ...
)
}
\arguments{
\item{filterable}{make columns filterable using filterable}

\item{searchable}{make the entire table searchable using searchable}

\item{defaultPageSize}{The page size options be customized 10 through pageSizeOptions}

\item{borderless}{customize table styling using Borderless}

\item{striped}{customize table styling using Stiped}

\item{highlight}{customize table styling using highlight}

\item{...}{}
}
\description{
Define default behavior of reactable
This function inputs a dataframe of observation for data,it will return data table use
reactable() on a data frame.
}
