% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dot}
\alias{plot_dot}
\title{Creating a dot plot}
\usage{
plot_dot(
  tbl,
  y_var,
  label,
  x_breaks = NULL,
  color = NULL,
  shape = NULL,
  title = "AE (\%)",
  background_color = c("#69B8F7", "#FFFFFF"),
  background_alpha = 0.3,
  theme = theme_panel(show_text = TRUE, show_ticks = TRUE),
  legend_nrow = 1
)
}
\arguments{
\item{tbl}{A data frame selected from input data set to display on this plot.
y and x variables are required.}

\item{y_var}{A character value, specifies a variable to be displayed on the
y-axis.}

\item{label}{A character vector of labels for each treatment group. The
control group label should be specified as the last element of the vector.}

\item{x_breaks}{A numeric vector for x-axis breaks. Default value is NULL,
which uses a default ggplot2 x-axis breaks presentation.}

\item{color}{Defines a color for each treatment group.}

\item{shape}{Defines a shape for each treatment group. Default is circle and
square. Input values could be either a character, e.g., "triangle" or numeric
value, e.g., 1. For details and examples see \code{vignette("example-color-shape-details", package = "mkvis")}.}

\item{title}{A panel title. Default is "AE (\%)".}

\item{background_color}{A plot background color. Default is c("#69B8F7",
"#FFFFFF"), which is pastel blue and white. The value of this argument is
used as input for \code{background_color} argument in \code{\link{background_panel}}.}

\item{background_alpha}{Controls the opacity of the background. Default is 0.3.
The value of this argument is the input for \code{background_alpha} argument in
\code{\link{background_panel}}.}

\item{theme}{Controls panel theme, including y-axis text, ticks and plot
margin. Default \code{theme_panel(show_text = TRUE, show_ticks = TRUE)}. For
more details refer to \code{\link{theme_panel}}.}

\item{legend_nrow}{An integer, sets a number of rows for a legend
display. Must be \code{<=} the number of the treatment groups. To omit the legend
set to NULL. Default is 1.}
}
\value{
AEs incidence plot by item and treatment group.
}
\description{
\code{plot_dot} creates a dot plot by item. For instance, could be
used to create AEs incidence plot by Preferred Term and treatment group,
as part of a rainfall plot.
}
\examples{
library(dplyr)
# Example 1: minimal example
forestly_adsl$TRTA <- factor(forestly_adsl$TRT01A, levels = c("Xanomeline Low Dose", "Placebo"), labels = c("Low Dose", "Placebo"))
forestly_adae$TRTA <- factor(forestly_adae$TRTA, levels = c("Xanomeline Low Dose", "Placebo"), labels = c("Low Dose", "Placebo"))
meta <- meta_forestly(
  dataset_adsl = forestly_adsl,
  dataset_adae = forestly_adae,
  population_term = "apat",
  observation_term = "wk12",
  parameter_term = "any;rel;ser"
) |>
  prepare_ae_forestly() |>
  format_ae_forestly()

meta_any <- meta$tbl[1:20,] |>
dplyr::filter(parameter == "any")
meta_any |>
dplyr::select(name, prop_1, prop_2) |>
plot_dot("name", label = c("Treatment", "Placebo"))
}
