% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_errorbar}
\alias{plot_errorbar}
\title{Creating plot to display risk difference}
\usage{
plot_errorbar(
  tbl,
  y_var,
  errbar_width = 0.4,
  color = NULL,
  shape = NULL,
  label,
  x_breaks = NULL,
  grp_abbrev = "paired",
  favor_direction = "negative",
  vline = NULL,
  line_type = 1,
  title = "Risk Diff. + 95\% CI \\n (Percentage Points)",
  background_color = c("#69B8F7", "#FFFFFF"),
  background_alpha = 0.3,
  theme = theme_panel(show_text = TRUE, show_ticks = TRUE),
  legend_nrow = 1
)
}
\arguments{
\item{tbl}{A data frame selected from input data set to display on this plot.
y and x variables are required.}

\item{y_var}{A character value, specifies a variable to be displayed on the
y-axis.}

\item{errbar_width}{A numeric value to define the error bar width. Default is
0.4. Value of this argument will be a half length of the error bar, e.g.,
\code{errorbar_width = 0.2} means half of the error bar width is 0.2 unit length.
If y = 4, the error bar will range from y = 3.8 to y = 4.2.}

\item{color}{Defines a color for each treatment group.}

\item{shape}{Defines a shape for each treatment group. Default is circle and
square. Input values could be either a character, e.g., "triangle" or numeric
value, e.g., 1. For details and examples see \code{vignette("example-color-shape-details", package = "mkvis")}.}

\item{label}{A character vector of labels for each treatment group. The
control group label should be specified as the last element of the vector.}

\item{x_breaks}{A numeric vector for x-axis breaks. Default value is NULL,
which uses a default ggplot2 x-axis breaks presentation.}

\item{grp_abbrev}{A character vector for displaying the treatment groups on a
favor bar. If grp_abbrev = "paired", treatment label on an error bar will be
same as in \code{label} argument. If grp_abbrev = "none", error bar will not be
shown. Also, for customized terms, a user can provide an alternative vector
of treatment labels. Default is "paired".}

\item{favor_direction}{Defines the position of a favor label under the
condition “comparison is treatment – control”. For AEs \code{favor_direction}
should be "negative"; for efficacy \code{favor_direction} should be "positive".}

\item{vline}{A vertical reference line position. Default is NULL. A user can
define one or multiple numeric values in a vector as a reference line
position.}

\item{line_type}{A reference line type. Default is solid line. A user can
define one or multiple line types in a vector (can be numeric such as 1, 2, 3
or character such as "solid", "dashed"). The value will be used recurrently
and order will be consistent with argument \code{vline}.}

\item{title}{A title of this plot. Default is "Risk Diff. + 95\% CI
\\n (Percentage Points)".}

\item{background_color}{A plot background color. Default is c("#69B8F7",
"#FFFFFF"), which is pastel blue and white. The value of this argument is
used as input for \code{background_color} argument in \code{\link{background_panel}}.}

\item{background_alpha}{Controls the opacity of the background. Default is 0.3.
The value of this argument is the input for \code{background_alpha} argument in
\code{\link{background_panel}}.}

\item{theme}{Controls panel theme, including y-axis text, ticks and plot
margin. Default \code{theme_panel(show_text = TRUE, show_ticks = TRUE)}. For
more details refer to \code{\link{theme_panel}}.}

\item{legend_nrow}{An integer, sets a number of rows for a legend
display. Must be \code{<=} the number of the treatment groups. To omit the legend
set to NULL. Default is 1.}
}
\value{
A risk difference plot for each item.
}
\description{
\code{plot_errorbar} creates a plot to display a risk difference for
each item
}
\examples{
library(dplyr)
# Example 1: minimal example
forestly_adsl$TRTA <- factor(forestly_adsl$TRT01A, levels = c("Xanomeline Low Dose", "Placebo"), labels = c("Low Dose", "Placebo"))
forestly_adae$TRTA <- factor(forestly_adae$TRTA, levels = c("Xanomeline Low Dose", "Placebo"), labels = c("Low Dose", "Placebo"))
meta <- meta_forestly(
  dataset_adsl = forestly_adsl,
  dataset_adae = forestly_adae,
  population_term = "apat",
  observation_term = "wk12",
  parameter_term = "any;rel;ser"
) |>
  prepare_ae_forestly() |>
  format_ae_forestly()

meta_any <- meta$tbl[1:20,] |>
dplyr::filter(parameter == "any")
meta_any |>
dplyr::select(name, diff_1, lower_1, upper_1) |>
plot_errorbar(y_var = "name",
              label = c("Treatment", "Placebo"))
}
