% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forest.R
\name{plot_forest}
\alias{plot_forest}
\title{Generate interactive forest plot with a select list as filter}
\usage{
plot_forest(
  db,
  fig_prop_color = c("#00857C", "#66203A"),
  fig_prop_label = NULL,
  fig_diff_color = "black",
  fig_diff_label = NULL,
  small_sample = NULL
)
}
\arguments{
\item{db}{a list of two components:
(1) db$table: ae data frame rbinded by interested ae labels
(2) db$listing: a standard adverse event data frame with select columns}

\item{fig_prop_color}{a character vector of length two to specify the colors to plot the two proportions.
The first entry refers to the color of the AE proportion in the treatment arm.
The second entry refers to the color of the AE proportion in the control arm.
The default value is c("gold", "purple").}

\item{fig_prop_label}{a character vector of length two to specify the x-axis label of the two propositions plots.
The first entry refers to the x-lab legend of the treatment arm.
The second entry refers to the x-lab legend of the control arm.
The default value is  c("treatment", "control")}

\item{fig_diff_color}{a string to specify the color to plot the error bar.
The default value is "black".}

\item{fig_diff_label}{a string to specify the x-axis legend of the error bar.
The default value is "treatment <- Favor -> control".}

\item{small_sample}{a integral vector of length 2. The first element is for treatment group and
the second element is for control group. The default value is c(4, 4).}
}
\value{
a reactable with a select list
}
\description{
Generate interactive forest plot with a select list as filter
}
\examples{
library(dplyr)
tb <- data.frame(ae = c("headache", "pain", "fever", "running nose", "fever", "headache", "running nose"),
                 ae_label = c("ALL", "ALL", "ALL", "ALL", "AESER", "AEREL", "AEREL"),
                 n_1 = c(40, 50, 30, 50, 30, 30, 50),
                 n_2 = c(20, 30, 10, 50, 10, 10, 30),
                 N_1 = rep(60, 7),
                 N_2 = rep(60, 7),
                 stratum = rep("NULL", 7))
tb <- tb \%>\% dplyr::mutate(pct_1 = n_1/N_1 * 100, pct_2 = n_2/N_2 * 100)
db <- list(table = tb, listing = tb \%>\% rename(AE = ae),
           sample_size = tibble::tibble(treatment = c("treatment", "control"), N = c(60, 60)),
           treatment_order = c("MK9999" = "Xanomeline", "Placebo" = "Placebo"))

plot_forest(db)
plot_forest(db, fig_prop_color = c("red", "green"), fig_diff_color = "blue")
}
