% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkline_draw_axis.R
\name{sparkline_draw_axis}
\alias{sparkline_draw_axis}
\title{sparkline_draw_axis}
\usage{
sparkline_draw_axis(
  color,
  label,
  xlim,
  xlab = NULL,
  mode = "markers",
  showlegend = TRUE,
  height = ifelse(showlegend, 60, 30),
  margin_bottom = ifelse(showlegend, 0, height)
)
}
\description{
sparkline_draw_axis
}
\examples{

color <- c("red", "blue")
label <- c("MK9999", "Placebo")
xlim  <- c(-2, 2)
xlab  <- "Example"

sparkline_draw_axis(color, label, xlim,
                    showlegend = FALSE, height = 30)

sparkline_draw_axis(color, label, xlim,
                    showlegend = TRUE, height = 60)

sparkline_draw_axis(color, label, xlim, xlab = "Example",
                    showlegend = FALSE, height = 60)

sparkline_draw_axis(color, label, xlim, xlab = "Example", 
                    showlegend = TRUE, height = 100)
                    
sparkline_draw_axis(color, label, xlim, xlab = "Example", 
                    mode = "lines", showlegend = TRUE, height = 100)
           
sparkline_draw_axis(color, label, xlim, xlab = "Example", 
                    mode = "markers+lines", showlegend = TRUE, height = 100)

}
