% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkline_errorbar.R
\name{sparkline_errorbar}
\alias{sparkline_errorbar}
\title{sparkline_errorbar}
\usage{
sparkline_errorbar(
  est,
  lower,
  upper,
  color = NULL,
  height = 30,
  xlim = NULL,
  vline = NULL,
  margin = list(l = 0, r = 0, b = 0, t = 0, pad = 0),
  ...
)
}
\description{
sparkline_errorbar
}
\examples{

est <- c(0,1,2)
lower <- c(-1, -2, -1)
upper <- c(3, 5, 5)
color <- c("red", "blue", "gold")
sparkline_errorbar(est, lower, upper, color)

}
