% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkline_point.R
\name{sparkline_point}
\alias{sparkline_point}
\title{sparkline_point}
\usage{
sparkline_point(
  x,
  y,
  color = NULL,
  height = 30,
  hover_text = NULL,
  xlim = NULL,
  vline = NULL,
  margin = list(l = 0, r = 0, b = 0, t = 0, pad = 0),
  ...
)
}
\description{
sparkline_point
}
\examples{

x <- 1:3
y <- 1:3
color <- c("red", "blue", "gold")
sparkline_point(x = x, y = y, color = color, height = 30)

}
