% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkline_point_js.R
\name{sparkline_point_js}
\alias{sparkline_point_js}
\title{Sparkline Point Figure Javascript Render Using Plotly}
\usage{
sparkline_point_js(
  tbl,
  x,
  x_lower = NULL,
  x_upper = x_lower,
  xlim = NULL,
  y = seq_along(x),
  vline = NULL,
  text = NULL,
  height = 30,
  width = 150,
  color = "gold",
  color_errorbar = color,
  color_vline = "#00000050"
)
}
\description{
Sparkline Point Figure Javascript Render Using Plotly
}
\examples{
library(reactable)
library(htmltools)

js <- sparkline_point_js(iris, "Sepal.Length")
js <- sparkline_point_js(iris, "Sepal.Length", text = "x")
js <- sparkline_point_js(iris, "Sepal.Length", text = "'count:' + x")
js <- sparkline_point_js(iris, "Sepal.Length", x_lower = "Sepal.Width", text = "'range:' + x + '(' + x_lower + ',' + x_upper + ')'" )
js <- sparkline_point_js(iris, "Sepal.Length", x_lower = "Sepal.Width")
js <- sparkline_point_js(iris, "Sepal.Length", vline = 6)
cat(js)

p <- reactable(data = iris[,1:2],
               theme = reactableTheme(cellPadding = "0px 8px"),
               borderless = TRUE,
               highlight = TRUE,
               resizable = TRUE,
               columns = list( Sepal.Length = colDef(html = TRUE,
                                                     cell = JS(js),
                                                     width = 150,
                                                     style="font-size: 0px; padding: 0px; margin: 0px;")))

browsable(tagList(
  reactR::html_dependency_react(),
  htmltools::tags$script(src="https://cdn.plot.ly/plotly-latest.min.js"),
  htmltools::tags$script(src="https://unpkg.com/react-plotly.js@1.0.2/dist/create-plotly-component.js"),
  p
))

}
