% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_ae_table.R
\name{tidy_ae_table}
\alias{tidy_ae_table}
\title{tidy_ae_table2}
\usage{
tidy_ae_table(
  population_from,
  observation_from,
  population_where = "ITTFL=='Y'",
  observation_where = NULL,
  treatment_var = "TRTA",
  treatment_order = unique(population_from[[treatment_var]]),
  ae_var = ae_var,
  ae_interested = ae_interested,
  stratum_var = NULL,
  listing_var = names(observation_from)
)
}
\arguments{
\item{population_from}{A data frame to obtain population level variables.Typically an 'adsl' dataset.}

\item{observation_from}{A data frame to obtain observation level variables.}

\item{population_where}{A character string to define the criteria to select analysis population.}

\item{observation_where}{A character string to define the criteria to select analysis observation.}

\item{treatment_var}{A character string to define the variable of new column called "treatment"}

\item{treatment_order}{A character vector to define the treatment display order and label.}

\item{ae_var}{A character string to define the variable of new column called ae}

\item{ae_interested}{A character string including the interested AE structure, ex. AESER, AEREL, ...}

\item{stratum_var}{A character string to define the variable of baseline stratum in 'population_from'.Only one 'stratum_var' is allowed.}

\item{listing_var}{A character string to define the criteria to select the column of the table}
}
\value{
Return a standard adverse event data frame
}
\description{
This function is used to obtain AE information ready for visualization
}
\examples{
library(dplyr)
db <- tidy_ae_table(population_from  = adsl \%>\% rename(TRTA = TRT01A),
                    observation_from = adae,
                    population_where = NULL,
                    observation_where = NULL,
                    treatment_var = "TRTA",
                    treatment_order = c("MK9999" = "Xanomeline High Dose", "Placebo" = "Placebo"),
                    ae_var = "AEDECOD",
                    ae_interested = ae_interested(ae_criterion = c('AESER == "Y"', 'AEREL != "NONE"'),
                                                  ae_label = c("with serious adverse events",
                                                               "with drug-related adverse events")),
                    listing_var = c("USUBJID", "SEX", "RACE", "AGE"))
}
